% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{AlpineStore}
\alias{AlpineStore}
\title{Alpine.js Store Manager}
\description{
Manages a synchronized state between R and an Alpine.js store.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Holds the values available in the store. Read and write like \code{reactiveValues()}
Initialize Instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlpineStore-new}{\code{AlpineStore$new()}}
\item \href{#method-AlpineStore-clone}{\code{AlpineStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$new(id, session = shiny::getDefaultReactiveDomain(), data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier for the store (used in $store.id)}

\item{\code{session}}{The Shiny session object}

\item{\code{data}}{Named list with keys and values to populate store on initialisation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
