% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{assert_plugin_active}
\alias{assert_plugin_active}
\title{Assert that a plugin is loaded}
\usage{
assert_plugin_active(plugin)
}
\arguments{
\item{plugin}{Character(1). The name of the required Alpine plugin}
}
\description{
Assert that a plugin is loaded
}
\keyword{internal}
