% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{assert_sherpa_template}
\alias{assert_sherpa_template}
\title{Sanity checks for Sherpa HTML Template}
\usage{
assert_sherpa_template(path)
}
\value{
The original path if valid, otherwise aborts.
}
\description{
Ensures \code{{{ headContent() }}} and \code{x-data} are present.
This ensures the template works correctly with Sherpa and Shiny
}
\keyword{internal}
