% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{html_template}
\alias{html_template}
\title{Create a Sherpa UI from an HTML Template}
\usage{
html_template(path, stores = NULL, plugins = NULL, ...)
}
\arguments{
\item{path}{Path to the .html file}

\item{stores}{Character vector of names of stores that will be used. Must be initialized here.}

\item{plugins}{Character vector with Alpine plugin names (e.g. c("persist", "mask", "intersect")).}

\item{...}{Additional variables to pass to the template (accessible via \code{{{ var }}})}
}
\description{
Create a Sherpa UI from an HTML Template
}
\seealso{
\code{\link[shiny:reexports]{shiny::htmlTemplate()}}
}
