% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{s}
\alias{s}
\title{The Sherpa Tag Proxy}
\format{
An object of class \code{environment} of length 181.
}
\usage{
s$div(...)
}
\description{
An environment containing wrappers for all shiny::tags.
These wrappers automatically handle sherpa attribute helpers
(like \code{\link[=x_data]{x_data()}}, \code{\link[=x_on]{x_on()}}, ...) without the need for the
 operator.
}
\details{
The sherpa attribute helpers all return a named list with
a single Element. By default, \link[shiny:reexports]{shiny::tags} will treat these named
lists as regular tag content instead of HTML-Attributes (which we want).

There are 3 ways to make \link[shiny:reexports]{shiny::tags} treat the sherpa attribute helpers
as HTML-attributes:
\itemize{
\item \code{splice} them into the function call using 
\item add them using \code{\link{\%s\%}}, a wrapper for \link[shiny:reexports]{shiny::tagAppendAttributes}
that automatically splices sherpa attribute helpers
\item using this "tag proxy". It proxies the \code{shiny::tags}-Object but adds
automatic splicing for sherpa attribute helpers.
}
}
\keyword{datasets}
