% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magics.R
\name{s_persist}
\alias{s_persist}
\title{Create $persist Magic}
\usage{
s_persist(value, as = NULL, using = NULL)
}
\arguments{
\item{value}{The initial value. This is automatically converted to JSON.}

\item{as}{Character. Optional: A custom key name for the storage.}

\item{using}{Character. Optional: The storage provider (e.g., 'sessionStorage').}
}
\description{
Helper to generate the \verb{$persist} magic for Alpine.js.
}
