% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magics.R
\name{s_store}
\alias{s_store}
\title{Access an Alpine Store}
\usage{
s_store(store_id, prop = NULL)
}
\arguments{
\item{store_id}{Unique ID of the store.}

\item{prop}{Optional property within the store.}
}
\value{
A string: \verb{$store.store_id.prop}.
}
\description{
Helper to generate the \verb{$store} syntax for Alpine stores.
}
