% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherpa-package.R
\docType{package}
\name{sherpa-package}
\alias{sherpa}
\alias{sherpa-package}
\title{sherpa: Carries your data between Shiny and Alpine.js}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sherpa is a bridge between Shiny and Alpine.js. It mirrors Shiny reactiveValues in an Alpine.js store in the browser and provides helpers for adding Alpine.js directives to Shiny UI tags.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Teebusch/sherpa}
  \item \url{https://teebusch.github.io/sherpa/}
  \item Report bugs at \url{https://github.com/Teebusch/sherpa/issues}
}

}
\author{
\strong{Maintainer}: Tobias Busch \email{teebusch@gmail.com} [copyright holder]

}
\keyword{internal}
