% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{\%s\%}
\alias{\%s\%}
\alias{sherpa_operator}
\title{Sherpa Attribute Operator}
\usage{
tag \%s\% ...
}
\arguments{
\item{tag}{A shiny tag}

\item{...}{Sherpa attributes (e.g., x_data(), x_cloak())}
}
\description{
Appends Sherpa attributes to a tag without needing the !!! operator.
}
