% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_json}
\alias{to_json}
\title{Convert R objects to JavaScript-ready strings}
\usage{
to_json(x)
}
\arguments{
\item{x}{The value to convert.}
}
\description{
Internal utility to ensure R types (logical, list, numeric)
translate correctly to Alpine.js expressions.
}
\keyword{internal}
