% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{watch}
\alias{watch}
\title{Activate Hot Module Reloading for a HTML-Template}
\usage{
watch(path, interval = 0.3, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{path}{Path to the HTML template to monitor.}

\item{interval}{How often to check for changes in seconds.}

\item{session}{The Shiny session object.}
}
\description{
Run within your server function to register a a watcher for a HTML-Template.
Usually the same template that is supplie to the ui via \code{\link[=html_template]{html_template()}}
}
