% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-plugins.R
\name{x_anchor}
\alias{x_anchor}
\title{Alpine.js Anchor Directive}
\usage{
x_anchor(ref, pivot = NULL, offset = NULL)
}
\arguments{
\item{ref}{The reference element (e.g., "$refs.button")}

\item{pivot}{The anchor point (e.g., "bottom-start", "top-end")}

\item{offset}{Offset in pixels}
}
\description{
Position an element relative to another "reference" element.
Requires Plugin "Anchor"
}
