% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_change}
\alias{x_change}
\title{Change Event Shortcut}
\usage{
x_change(code, modifiers = NULL)
}
\arguments{
\item{code}{JavaScript code.}

\item{modifiers}{Event modifiers.}
}
\description{
Shortcut for \code{x-on:change}.
}
