% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_click}
\alias{x_click}
\title{Click Event Shortcut}
\usage{
x_click(code, modifiers = NULL)
}
\arguments{
\item{code}{JavaScript code.}

\item{modifiers}{Event modifiers like "prevent" or "stop".}
}
\description{
Shortcut for \code{x-on:click}.
}
