% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_cloak}
\alias{x_cloak}
\title{Hide Elements until Alpine is Loaded}
\usage{
x_cloak()
}
\description{
Prevents "flicker" by hiding elements before Alpine initializes.
Requires CSS: \verb{[x-cloak] \{ display: none !important; \}}
}
\seealso{
\url{https://alpinejs.dev/directives/cloak}
}
