% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_collapse}
\alias{x_collapse}
\title{Alpine.js Collapse Directive}
\usage{
x_collapse(duration = NULL, min = NULL)
}
\arguments{
\item{duration}{Optional duration (e.g., "500ms")}

\item{min}{Optional minimum height (e.g., "50px")}
}
\description{
Expand and collapse elements using smooth animations.
Requires Plugin "Collapse"
}
\seealso{
\url{https://alpinejs.dev/plugins/collapse}
}
