% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_data}
\alias{x_data}
\title{Define Alpine Component Data}
\usage{
x_data(data = NA)
}
\arguments{
\item{data}{A JSON string (e.g., \code{"{ open: false }"}) or a store name.
Use \code{NA} to create a data-less component.}
}
\description{
Provides reactive data for an Alpine component.
}
\seealso{
\url{https://alpinejs.dev/directives/data}
}
