% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_effect}
\alias{x_effect}
\title{Reactive Effects (x-effect)}
\usage{
x_effect(code)
}
\arguments{
\item{code}{JavaScript code.}
}
\description{
Run a script when a reactive dependency changes.
}
\seealso{
\url{https://alpinejs.dev/directives/effect}
}
