% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_for}
\alias{x_for}
\title{Loop Through Data (x-for)}
\usage{
x_for(iteration, key_prop = NULL)
}
\arguments{
\item{iteration}{Loop syntax, e.g., "item in items".}

\item{key_prop}{Optional property name to use as a unique key.}
}
\description{
Elements inside an x-for must be wrapped in a \verb{<template>} tag.
}
\seealso{
\url{https://alpinejs.dev/directives/for}
}
