% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_html}
\alias{x_html}
\title{Bind HTML Content}
\usage{
x_html(expression)
}
\arguments{
\item{expression}{JavaScript expression returning an HTML string.}
}
\description{
Update the inner HTML of an element.
\strong{Warning:} Only use with trusted content to prevent XSS attacks.
}
\seealso{
\url{https://alpinejs.dev/directives/html}
}
