% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_id}
\alias{x_id}
\title{Generate Unique IDs}
\usage{
x_id(id_name)
}
\arguments{
\item{id_name}{The name/key for the unique ID.}
}
\description{
Create unique IDs that can be referenced by multiple elements (e.g., aria-labelledby).
}
\seealso{
\url{https://alpinejs.dev/directives/id}
}
