% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_if}
\alias{x_if}
\title{Conditional Rendering (x-if)}
\usage{
x_if(condition)
}
\arguments{
\item{condition}{JavaScript expression.}
}
\description{
Elements inside an x-if must be wrapped in a \verb{<template>} tag.
}
\seealso{
\url{https://alpinejs.dev/directives/if}
}
