% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_ignore}
\alias{x_ignore}
\title{Ignore Element for Alpine}
\usage{
x_ignore()
}
\description{
Prevents Alpine from initializing or looking for directives inside an element.
Useful for integrating other JS libraries (like Leaflet or D3).
}
\seealso{
\url{https://alpinejs.dev/directives/ignore}
}
