% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_init}
\alias{x_init}
\title{Alpine Initialization Hook}
\usage{
x_init(code)
}
\arguments{
\item{code}{JavaScript code string.}
}
\description{
Execute JavaScript code during the initialization phase of an element.
}
\seealso{
\url{https://alpinejs.dev/directives/init}
}
