% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-plugins.R
\name{x_intersect}
\alias{x_intersect}
\title{Alpine.js Intersect Directive}
\usage{
x_intersect(expression, once = FALSE)
}
\arguments{
\item{expression}{The JS code to run on intersection}

\item{once}{Logical, if TRUE, only trigger the first time}
}
\description{
React when an element enters the viewport.
Requires Plugin "Intersect"
}
