% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes-plugins.R
\name{x_mask}
\alias{x_mask}
\title{Alpine.js Mask Directive}
\usage{
x_mask(mask, quote = TRUE)
}
\arguments{
\item{mask}{The mask pattern (e.g., "99/99/9999")}

\item{quote}{Logical, if TRUE, wraps the mask in single quotes (needed for strings)}
}
\description{
Automatically format input fields as a user types.
Requires Plugin "Mask"
}
