% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_model}
\alias{x_model}
\title{Two-Way Data Binding}
\usage{
x_model(prop, modifiers = NULL)
}
\arguments{
\item{prop}{The property in the Alpine data object.}

\item{modifiers}{Modifiers like "debounce.500ms", "number", or "lazy".}
}
\description{
Bind the value of an input element to Alpine data.
}
\seealso{
\url{https://alpinejs.dev/directives/model}
}
