% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_modelable}
\alias{x_modelable}
\title{Expose Properties to x-model}
\usage{
x_modelable(prop)
}
\arguments{
\item{prop}{The property to make modelable.}
}
\description{
Make any Alpine property available as the target for \code{x-model}.
}
\seealso{
\url{https://alpinejs.dev/directives/modelable}
}
