% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_on}
\alias{x_on}
\title{Event Listener}
\usage{
x_on(event, code, modifiers = NULL)
}
\arguments{
\item{event}{The event name (e.g., "click", "keyup", "submit").}

\item{code}{JavaScript code to execute.}

\item{modifiers}{Optional modifiers like "prevent", "stop", or "outside".}
}
\description{
Listen for browser events and execute JavaScript code.
}
\seealso{
\url{https://alpinejs.dev/directives/on}
}
