% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_ref}
\alias{x_ref}
\title{Utility for accessing DOM elements directly.}
\usage{
x_ref(ref)
}
\arguments{
\item{ref}{The reference to give the DOM-element.}
}
\description{
It's most useful as a replacement for APIs like getElementById and querySelector.
}
\seealso{
\url{https://alpinejs.dev/directives/ref}
}
