% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_resize}
\alias{x_resize}
\title{Alpine.js Resize Directive}
\usage{
x_resize(expression)
}
\arguments{
\item{expression}{The JS code to run when resized}
}
\description{
Execute code whenever an element's size changes.
Requires Plugin "Resize"
}
\seealso{
\url{https://alpinejs.dev/plugins/resize}
}
