% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_show}
\alias{x_show}
\title{Toggle Visibility}
\usage{
x_show(condition)
}
\arguments{
\item{condition}{JavaScript expression returning a boolean-like value.}
}
\description{
Show and hide DOM elements via the 'display: none' CSS property.
}
\seealso{
\url{https://alpinejs.dev/directives/show}
}
