% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_store_ref}
\alias{x_store_ref}
\title{Access an Alpine Store}
\usage{
x_store_ref(store_id, prop = NULL)
}
\arguments{
\item{store_id}{Unique ID of the store.}

\item{prop}{Optional property within the store.}
}
\value{
A string: \verb{$store.store_id.prop}.
}
\description{
Helper to generate the \verb{$store} syntax for Alpine stores.
}
