% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_teleport}
\alias{x_teleport}
\title{Teleport DOM Elements}
\usage{
x_teleport(selector)
}
\arguments{
\item{selector}{CSS selector for the target destination (e.g., "body").}
}
\description{
Render an element's content in a different part of the DOM.
}
\seealso{
\url{https://alpinejs.dev/directives/teleport}
}
