% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_text}
\alias{x_text}
\title{Bind Text Content}
\usage{
x_text(expression)
}
\arguments{
\item{expression}{JavaScript expression returning a string.}
}
\description{
Update the text content of an element automatically.
}
\seealso{
\url{https://alpinejs.dev/directives/text}
}
