% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_transition}
\alias{x_transition}
\title{Apply CSS Transitions}
\usage{
x_transition(modifiers = NULL)
}
\arguments{
\item{modifiers}{Optional modifiers for specific stages (e.g., "enter", "leave", "duration.500ms").}
}
\description{
Smoothly transition elements as they enter/leave the DOM.
}
\seealso{
\url{https://alpinejs.dev/directives/transition}
}
