% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_trap}
\alias{x_trap}
\title{Alpine.js Focus Trap Directive}
\usage{
x_trap(expression, inertia = FALSE, noscroll = FALSE)
}
\arguments{
\item{expression}{A JS expression evaluating to boolean (e.g., "isOpen")}

\item{inertia}{Logical, if TRUE, allows focus to move to browser chrome}

\item{noscroll}{Logical, if TRUE, prevents scrolling when focus is trapped}
}
\description{
Trap focus within a specific DOM element (perfect for modals).
Requires Plugin "Focus"
}
\seealso{
\url{https://alpinejs.dev/plugins/focus}
}
