% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{AlpineStore}
\alias{AlpineStore}
\title{Alpine.js Store Manager}
\description{
Alpine.js Store Manager

Alpine.js Store Manager
}
\details{
This class manages a synchronized state between R and an Alpine.js store.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The store's ID (will be the same in the Client)}

\item{\code{session}}{The Shiny session}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlpineStore-new}{\code{AlpineStore$new()}}
\item \href{#method-AlpineStore-sync}{\code{AlpineStore$sync()}}
\item \href{#method-AlpineStore-update}{\code{AlpineStore$update()}}
\item \href{#method-AlpineStore-init}{\code{AlpineStore$init()}}
\item \href{#method-AlpineStore-clone}{\code{AlpineStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Alpine Store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$new(id, session = shiny::getDefaultReactiveDomain())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier for the store (used in $store.id)}

\item{\code{session}}{The Shiny session object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-sync"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-sync}{}}}
\subsection{Method \code{sync()}}{
Bind a reactive expression to a store property
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$sync(name, reactive_expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the property in the Alpine store}

\item{\code{reactive_expr}}{A reactive expression or a value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-update"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-update}{}}}
\subsection{Method \code{update()}}{
Send a direct update to the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$update(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Property name}

\item{\code{value}}{The value to send (will be converted to JSON)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-init"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-init}{}}}
\subsection{Method \code{init()}}{
Initialize the store with a full set of data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$init(data = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A named list of initial values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlpineStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlpineStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlpineStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
