% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherpa-package.R
\docType{package}
\name{sherpa-package}
\alias{sherpa}
\alias{sherpa-package}
\title{sherpa: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Teebusch/sherpa}
  \item \url{https://teebusch.github.io/sherpa/}
  \item Report bugs at \url{https://github.com/Teebusch/sherpa/issues}
}

}
\author{
\strong{Maintainer}: Tobias Busch \email{teebusch@gmail.com}

}
\keyword{internal}
