% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherpa-package.R
\docType{package}
\name{sherpa-package}
\alias{sherpa}
\alias{sherpa-package}
\title{sherpa: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tobias Busch

}
\keyword{internal}
