% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherpa-package.R
\docType{package}
\name{sherpa-package}
\alias{sherpa}
\alias{sherpa-package}
\title{sherpa: A lightweight bridge between R Shiny and Alpine.js}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sherpa allows you to build highly reactive user interfaces by handling state and transitions in the browser while keeping your business logic in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Teebusch/sherpa}
  \item \url{https://teebusch.github.io/sherpa/}
  \item Report bugs at \url{https://github.com/Teebusch/sherpa/issues}
}

}
\author{
\strong{Maintainer}: Tobias Busch \email{teebusch@gmail.com}

}
\keyword{internal}
