% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{use_alpine}
\alias{use_alpine}
\title{Activate Alpine.js and plugins in Shiny app}
\usage{
use_alpine(plugins = NULL, stores = NULL)
}
\arguments{
\item{plugins}{Character vector with Alpine plugin names (e.g. c("persist", "mask", "intersect")).}

\item{stores}{Character vector of names of stores that will be used. Must be initialized here.}
}
\value{
A \code{htmltools::tagList} with the dependencies as \link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Add this anywhere in your UI. Shiny will take care of hoisting it into your HTML page's \verb{<head>}.
}
\examples{
\dontrun{
ui <- fluidPage(
  use_alpine(plugins = c("mask", "persist")),
  div(x_data("{ open: false }"), ...)
)
}
}
