% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{x_bind}
\alias{x_bind}
\title{Attribute Binding}
\usage{
x_bind(attr, expression, modifiers = NULL)
}
\arguments{
\item{attr}{The HTML attribute to bind (e.g., "class", "disabled", "href").}

\item{expression}{JavaScript expression to evaluate.}

\item{modifiers}{Optional modifiers, such as "camel" for SVG attributes.}
}
\description{
Bind HTML attributes based on JavaScript expressions.
}
\seealso{
\url{https://alpinejs.dev/directives/bind}
}
