% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directives.R
\name{x_sort}
\alias{x_sort}
\title{Alpine.js Sort Directive}
\usage{
x_sort(expression = NA, group = NULL, handle = NULL)
}
\arguments{
\item{expression}{Optional JS code to run when sorting (e.g. to update a store)}

\item{group}{Optional group name to allow sorting between multiple lists}

\item{handle}{Optional selector for a drag handle (e.g., '.drag-handle')}
}
\description{
Re-order elements by dragging them with your mouse.
Requires Plugin "Sort"
}
\seealso{
\url{https://alpinejs.dev/plugins/sort}
}
